Findxo/django_api/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
PyJWT>=2.8
stripe>=9.0,<10.0